// app/src/generated/models/Csa_projectsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Csa_projectsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Csa_projectsstatecode = keyof typeof Csa_projectsstatecode;
export const Csa_projectsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Csa_projectsstatuscode = keyof typeof Csa_projectsstatuscode;

export interface Csa_projectsBase {
  csa_active?: boolean;
  "csa_CustomerId@odata.bind": string;
  csa_name?: string;
  csa_projectid: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Csa_projectsstatecode;
  statuscode?: Csa_projectsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Csa_projects extends Csa_projectsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  csa_activename?: string;
  csa_customeridname?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  csa_customerid?: object;
  _csa_customerid_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}